Fix test concurrency issue

+++ tst/CTA.Rules.Test/CTATests.cs
@@ -11,6 +11,7 @@ using System.IO.Compression;
 using System.Linq;
 using System.Reflection;
 using System.Text;
+using System.Threading;
 
 namespace CTA.Rules.Test
 {
@@ -252,7 +253,23 @@ namespace CTA.Rules.Test
         [TearDown]
         public void Cleanup()
         {
-            Directory.Delete(tempDir, true);
+            DeleteDir(0);
+        }
+
+        private void DeleteDir(int retries)
+        {
+            if (retries <= 10)
+            {
+                try
+                {
+                    Directory.Delete(tempDir, true);
+                }
+                catch (Exception)
+                {
+                    Thread.Sleep(60000);
+                    DeleteDir(retries + 1);
+                }
+            }
         }
     }
 }
